rename and move files to ovmm sub folders
