+++ habitat-lab/habitat/config/benchmark/ovmm/gaze.yaml
@@ -2,46 +2,36 @@
 
 defaults:
   - /habitat: habitat_config_base
-  - /habitat/simulator/agents@habitat.simulator.agents.main_agent: rgbdp_head_rgb_third_agent
-  - /habitat/task/rearrange: ovmm
+  - /habitat/simulator/agents@habitat.simulator.agents.main_agent: depthp_head_rgb_third_agent
+  - /habitat/task/ovmm: gaze
   - /habitat/dataset/rearrangement: hssd
   - _self_
 
 habitat:
-  seed: 7
   gym:
-    auto_name: NavToObj
+    auto_name: Pick
     obs_keys:
       - robot_head_depth
-      - robot_head_rgb
-      - robot_third_rgb
       - start_receptacle
-      - goal_receptacle
-      - robot_start_gps
-      - robot_start_compass
-      - joint
       - object_segmentation
       - start_recep_segmentation
-      - goal_recep_segmentation
-      - object_category
-      - camera_pose
       - object_embedding
-      - is_holding
-      - relative_resting_position
-      - ovmm_nav_goal_segmentation
-      - receptacle_segmentation
+      - joint
   environment:
-    max_episode_steps: 1250
+    max_episode_steps: 300
   simulator:
-    type: ObjectRearrangeSim-v0
+    type: OVMMSim-v0
     additional_object_paths:
       - data/objects/amazon_berkeley/configs/
       - data/objects/google_scanned/configs/
       - data/objects/ai2thorhab/configs/objects/
       - data/objects/floorplanner/configs/objects/
     debug_render_goal: False
+    debug_render: False
+    needs_markers: False
     concur_render: True
     auto_sleep: True
+    requires_textures: False
     kinematic_mode: True
     agents:
       main_agent:
@@ -49,25 +39,17 @@ habitat:
         height: 1.41
         robot_urdf: data/robots/hab_stretch/urdf/hab_stretch.urdf
         robot_type: "StretchRobot"
-        ik_arm_urdf: null
         sim_sensors:
-           head_rgb_sensor:
-             height: 640
-             width: 480
-             hfov: 42
            head_depth_sensor:
-             height: 640
-             width: 480
+             height: 160
+             width: 120
              hfov: 42
+             position: [0, 1.31, 0]
            head_panoptic_sensor:
-             height: 640
-             width: 480
+             height: 160
+             width: 120
              hfov: 42
+             position: [0, 1.31, 0]
     habitat_sim_v0:
       allow_sliding: False
       enable_physics: True
-    needs_markers: false
-  dataset:
-    data_path: data/episodes/rearrange/v4/val/cat_npz-exp-with_init_poses.json.gz
-    split: val
-
