+++ habitat-lab/habitat/config/benchmark/ovmm/nav_to_obj.yaml
@@ -2,33 +2,37 @@
 
 defaults:
   - /habitat: habitat_config_base
-  - /habitat/simulator/agents@habitat.simulator.agents.main_agent: rgbdp_head_rgb_third_agent
-  - /habitat/task/rearrange: cat_pick
+  - /habitat/simulator/agents@habitat.simulator.agents.main_agent: depthp_head_rgb_third_agent
+  - /habitat/task/ovmm: nav_to_obj
   - /habitat/dataset/rearrangement: hssd
   - _self_
 
 habitat:
   gym:
-    auto_name: Pick
+    auto_name: NavToObj
     obs_keys:
-      - robot_head_rgb
       - robot_head_depth
       - object_embedding
-      - joint
-      - is_holding
-      - relative_resting_position
+      - ovmm_nav_goal_segmentation
+      - receptacle_segmentation
+      - start_receptacle
+      - robot_start_gps
+      - robot_start_compass
   environment:
-    max_episode_steps: 300
+    max_episode_steps: 400
   simulator:
-    type: ObjectRearrangeSim-v0
+    type: OVMMSim-v0
     additional_object_paths:
       - data/objects/amazon_berkeley/configs/
       - data/objects/google_scanned/configs/
       - data/objects/ai2thorhab/configs/objects/
       - data/objects/floorplanner/configs/objects/
-    needs_markers: False
+    debug_render_goal: False
+    debug_render: False
     concur_render: True
     auto_sleep: True
+    requires_textures: False
+    kinematic_mode: True
     agents:
       main_agent:
         radius: 0.3
@@ -36,21 +40,17 @@ habitat:
         robot_urdf: data/robots/hab_stretch/urdf/hab_stretch.urdf
         robot_type: "StretchRobot"
         sim_sensors:
-           head_rgb_sensor:
-             height: 640
-             width: 480
-             hfov: 42
-             position: [0,1.31,0]
            head_depth_sensor:
-             height: 640
-             width: 480
+             height: 160
+             width: 120
              hfov: 42
              position: [0,1.31,0]
            head_panoptic_sensor:
-             height: 640
-             width: 480
+             height: 160
+             width: 120
              hfov: 42
              position: [0,1.31,0]
     habitat_sim_v0:
       allow_sliding: False
       enable_physics: True
+    needs_markers: false
