+++ habitat-lab/habitat/config/benchmark/ovmm/nav_to_rec.yaml
@@ -3,30 +3,29 @@
 defaults:
   - /habitat: habitat_config_base
   - /habitat/simulator/agents@habitat.simulator.agents.main_agent: depthp_head_rgb_third_agent
-  - /habitat/task/rearrange: static_place
+  - /habitat/task/ovmm: nav_to_rec
   - /habitat/dataset/rearrangement: hssd
   - _self_
 
 habitat:
   gym:
-    auto_name: StaticPlace
+    auto_name: NavToObj
     obs_keys:
       - robot_head_depth
-      - goal_receptacle
-      - object_embedding
       - ovmm_nav_goal_segmentation
-      - joint
-      - is_holding
+      - receptacle_segmentation
+      - goal_receptacle
+      - robot_start_gps
+      - robot_start_compass
   environment:
-    max_episode_steps: 350
+    max_episode_steps: 400
   simulator:
-    type: ObjectRearrangeSim-v0
+    type: OVMMSim-v0
     additional_object_paths:
       - data/objects/amazon_berkeley/configs/
       - data/objects/google_scanned/configs/
       - data/objects/ai2thorhab/configs/objects/
       - data/objects/floorplanner/configs/objects/
-    needs_markers: False
     debug_render_goal: False
     debug_render: False
     concur_render: True
@@ -53,3 +52,4 @@ habitat:
     habitat_sim_v0:
       allow_sliding: False
       enable_physics: True
+    needs_markers: false
