--- habitat-lab/habitat/config/habitat/dataset/rearrangement/floorplanner.yaml
@@ -1,12 +1,0 @@
-# @package habitat.dataset
-
-defaults:
-  - /habitat/dataset: object_rearrange_dataset_config_schema
-  - _self_
-
-type: RearrangeDataset-v0
-split: minitrain
-data_path: data/episodes/rearrange/v2/{split}/cat_npz-exp.json.gz
-scenes_dir: "data/fpss/fphab"
-viewpoints_matrix_path: data/episodes/rearrange/v2/{split}/cat_npz-exp-vps.npy
-transformations_matrix_path: data/episodes/rearrange/v2/{split}/cat_npz-exp-trans.npy
