+++ habitat-lab/habitat/config/habitat/dataset/rearrangement/hssd.yaml
@@ -4,7 +4,7 @@ defaults:
   - /habitat/dataset: object_rearrange_dataset_config_schema
   - _self_
 
-type: ObjectRearrangeDataset-v0
+type: OVMMDataset-v0
 split: train
 data_path: data/episodes/rearrange/v4/{split}/cat_npz-exp.json.gz
 scenes_dir: "data/fphab/fphab/"
