+++ habitat-lab/habitat/config/habitat/task/ovmm/gaze.yaml
