+++ habitat-lab/habitat/config/habitat/task/ovmm/nav_to_rec.yaml
@@ -22,15 +22,14 @@ defaults:
   - /habitat/task/lab_sensors:
     - joint_sensor
     - joint_velocity_sensor
-    - start_receptacle_sensor
+    - goal_receptacle_sensor
     - ovmm_nav_goal_segmentation_sensor
     - receptacle_segmentation_sensor
-    - object_embedding_sensor
     - robot_start_gps_sensor
     - robot_start_compass_sensor
   - _self_
 
-type: CatNavToObjTask-v0
+type: OVMMNavToObjTask-v0
 reward_measure: ovmm_nav_to_obj_reward
 success_measure: ovmm_nav_to_obj_success
 success_reward: 10.0
@@ -38,7 +37,7 @@ slack_reward: -0.005
 end_on_success: True
 constraint_violation_ends_episode: False
 constraint_violation_drops_object: True
-object_in_hand_sample_prob: 0.0
+object_in_hand_sample_prob: 1.0
 actions:
   base_velocity:
     type: "BaseWaypointTeleportAction"
