+++ habitat-lab/habitat/config/habitat/task/ovmm/ovmm.yaml
@@ -57,7 +57,7 @@ defaults:
     - receptacle_segmentation_sensor
   - _self_
 
-type: CatNavToObjTask-v0
+type: OVMMNavToObjTask-v0
 reward_measure: ovmm_place_success
 success_measure: ovmm_place_success
 success_reward: 10.0
