--- habitat-lab/habitat/config/habitat/task/rearrange/cat_pick.yaml
@@ -1,74 +1,0 @@
-# @package habitat.task
-
-defaults:
-  - /habitat/task: task_config_base
-  - /habitat/task/actions:
-    - arm_action
-    - base_velocity
-  - /habitat/task/measurements:
-    - robot_force
-    - force_terminate
-    - robot_colls
-    - end_effector_to_rest_distance
-    - end_effector_to_object_distance
-    - did_pick_object
-    - pick_distance_to_goal
-    - pick_distance_to_goal_reward
-    - pick_success
-    - pick_reward
-    - did_violate_hold_constraint
-    - num_steps
-  - /habitat/task/lab_sensors:
-    - joint_sensor
-    - joint_velocity_sensor
-    - is_holding_sensor
-    - end_effector_sensor
-    - relative_resting_pos_sensor
-    - object_category_sensor
-    - object_segmentation_sensor
-    - object_embedding_sensor
-  - _self_
-
-type: RearrangePickTask-v0
-base_angle_noise: 0.0
-constraint_violation_ends_episode: False
-constraint_violation_drops_object: True
-reward_measure: "pick_reward"
-success_measure: "pick_success"
-success_reward: 10.0
-slack_reward: -0.005
-end_on_success: True
-desired_resting_position: [0.119, -0.383, 0.657]
-start_in_manip_mode: True
-actions:
-  arm_action:
-    type: "ArmAction"
-    arm_controller: "ArmRelPosReducedActionStretch"
-    grip_controller: "MagicGraspAction"
-    arm_joint_mask: [1,0,0,0,1,1,1,1,0,0]
-    arm_joint_dimensionality: 5
-    grasp_thresh_dist: 0.15
-    disable_grip: False
-    max_delta_pos: 0.0125
-    ee_ctrl_lim: 0.015
-  base_velocity:
-    type: "BaseWaypointTeleportAction"
-    navmesh_offset: [[0.0, 0.0]]
-
-measurements:
-  force_terminate:
-    max_accum_force: 10_000.0
-    max_instant_force: 10_000.0
-  pick_distance_to_goal:
-    distance_to: EUCLIDEAN_POINT
-    distance_from: END_EFFECTOR
-    goals_attr: candidate_objects
-  pick_success:
-    object_goal: True
-  pick_reward:
-    object_goal: True
-lab_sensors:
-  joint_sensor:
-    dimensionality: 10
-  joint_velocity_sensor:
-    dimensionality: 10
