+++ habitat-lab/habitat/tasks/rearrange/rearrange_sensors.py
@@ -4,7 +4,6 @@
 # This source code is licensed under the MIT license found in the
 # LICENSE file in the root directory of this source tree.
 
-import pickle
 from typing import Any, Optional
 
 import numpy as np
@@ -13,10 +12,7 @@ from gym import spaces
 from habitat.core.embodied_task import Measure
 from habitat.core.registry import registry
 from habitat.core.simulator import Sensor, SensorTypes
-from habitat.datasets.rearrange.rearrange_dataset import (
-    ObjectRearrangeDatasetV0,
-    ObjectRearrangeEpisode,
-)
+
 from habitat.tasks.nav.nav import PointGoalSensor
 from habitat.tasks.rearrange.rearrange_sim import RearrangeSim
 from habitat.tasks.rearrange.utils import (
@@ -28,238 +24,6 @@ from habitat.tasks.rearrange.utils import (
 )
 from habitat.tasks.utils import cartesian_to_polar
 
-
-@registry.register_sensor
-class ObjectCategorySensor(Sensor):
-    cls_uuid: str = "object_category"
-
-    def __init__(
-        self,
-        sim,
-        config,
-        dataset: "ObjectRearrangeDatasetV0",
-        category_attribute="object_category",
-        name_to_id_mapping="obj_category_to_obj_category_id",
-        *args: Any,
-        **kwargs: Any,
-    ):
-        self._sim = sim
-        self._dataset = dataset
-        self._category_attribute = category_attribute
-        self._name_to_id_mapping = name_to_id_mapping
-
-        super().__init__(config=config)
-
-    def _get_uuid(self, *args: Any, **kwargs: Any) -> str:
-        return self.cls_uuid
-
-    def _get_sensor_type(self, *args: Any, **kwargs: Any):
-        return SensorTypes.SEMANTIC
-
-    def _get_observation_space(self, *args: Any, **kwargs: Any):
-        sensor_shape = (1,)
-        max_value = max(
-            getattr(self._dataset, self._name_to_id_mapping).values()
-        )
-
-        return spaces.Box(
-            low=0, high=max_value, shape=sensor_shape, dtype=np.int64
-        )
-
-    def get_observation(
-        self,
-        observations,
-        *args: Any,
-        episode: ObjectRearrangeEpisode,
-        **kwargs: Any,
-    ) -> Optional[np.ndarray]:
-        category_name = getattr(episode, self._category_attribute)
-        return np.array(
-            [getattr(self._dataset, self._name_to_id_mapping)[category_name]],
-            dtype=np.int64,
-        )
-
-
-@registry.register_sensor
-class ObjectEmbeddingSensor(Sensor):
-    cls_uuid: str = "object_embedding"
-
-    def __init__(
-        self,
-        sim,
-        config,
-        *args: Any,
-        **kwargs: Any,
-    ):
-        self._config = config
-        self._dimensionality = self._config.dimensionality
-        with open(config.embeddings_file, "rb") as f:
-            self._embeddings = pickle.load(f)
-
-        super().__init__(config=config)
-
-    def _get_uuid(self, *args: Any, **kwargs: Any) -> str:
-        return self.cls_uuid
-
-    def _get_sensor_type(self, *args: Any, **kwargs: Any):
-        return SensorTypes.TENSOR
-
-    def _get_observation_space(self, *args, **kwargs):
-        return spaces.Box(
-            shape=(self._dimensionality,),
-            low=np.finfo(np.float32).min,
-            high=np.finfo(np.float32).max,
-            dtype=np.float32,
-        )
-
-    def get_observation(self, observations, *args, episode, **kwargs):
-        category_name = episode.object_category
-        return self._embeddings[category_name]
-
-
-@registry.register_sensor
-class GoalReceptacleSensor(ObjectCategorySensor):
-    cls_uuid: str = "goal_receptacle"
-
-    def __init__(
-        self,
-        sim,
-        config,
-        dataset: "ObjectRearrangeDatasetV0",
-        *args: Any,
-        **kwargs: Any,
-    ):
-        super().__init__(
-            sim=sim,
-            config=config,
-            dataset=dataset,
-            category_attribute="goal_recep_category",
-            name_to_id_mapping="recep_category_to_recep_category_id",
-        )
-
-
-@registry.register_sensor
-class StartReceptacleSensor(ObjectCategorySensor):
-    cls_uuid: str = "start_receptacle"
-
-    def __init__(
-        self,
-        sim,
-        config,
-        dataset: "ObjectRearrangeDatasetV0",
-        *args: Any,
-        **kwargs: Any,
-    ):
-        super().__init__(
-            sim=sim,
-            config=config,
-            dataset=dataset,
-            category_attribute="start_recep_category",
-            name_to_id_mapping="recep_category_to_recep_category_id",
-        )
-
-
-@registry.register_sensor
-class ObjectSegmentationSensor(Sensor):
-    cls_uuid: str = "object_segmentation"
-    panoptic_uuid: str = "robot_head_panoptic"
-
-    def __init__(
-        self,
-        sim,
-        config,
-        *args: Any,
-        **kwargs: Any,
-    ):
-        self._config = config
-        self._blank_out_prob = self._config.blank_out_prob
-        self._sim = sim
-        self._instance_ids_start = self._sim.habitat_config.instance_ids_start
-        self._resolution = (
-            sim.agents[0]
-            ._sensors[self.panoptic_uuid]
-            .specification()
-            .resolution
-        )
-
-        super().__init__(config=config)
-
-    def _get_uuid(self, *args: Any, **kwargs: Any) -> str:
-        return self.cls_uuid
-
-    def _get_sensor_type(self, *args: Any, **kwargs: Any):
-        return SensorTypes.TENSOR
-
-    def _get_observation_space(self, *args, **kwargs):
-        return spaces.Box(
-            shape=(
-                self._resolution[0],
-                self._resolution[1],
-                1,
-            ),
-            low=0,
-            high=1,
-            dtype=np.uint8,
-        )
-
-    def get_observation(self, observations, *args, episode, task, **kwargs):
-        if np.random.random() < self._blank_out_prob:
-            return np.zeros_like(
-                observations[self.panoptic_uuid], dtype=np.uint8
-            )
-        else:
-            segmentation_sensor = np.zeros_like(
-                observations[self.panoptic_uuid], dtype=np.uint8
-            )
-            for g in episode.candidate_objects_hard:
-                segmentation_sensor = segmentation_sensor | (
-                    observations[self.panoptic_uuid]
-                    == self._sim.scene_obj_ids[int(g.object_id)]
-                    + self._instance_ids_start
-                )
-            return segmentation_sensor
-
-
-@registry.register_sensor
-class RecepSegmentationSensor(ObjectSegmentationSensor):
-    cls_uuid: str = "recep_segmentation"
-
-    def _get_recep_goals(self, episode):
-        raise NotImplementedError
-
-    def get_observation(self, observations, *args, episode, task, **kwargs):
-        recep_goals = self._get_recep_goals(episode)
-        if np.random.random() < self._config.blank_out_prob:
-            return np.zeros_like(
-                observations[self.panoptic_uuid], dtype=np.uint8
-            )
-        else:
-            segmentation_sensor = np.zeros_like(
-                observations[self.panoptic_uuid], dtype=np.uint8
-            )
-            for g in recep_goals:
-                segmentation_sensor = segmentation_sensor | (
-                    observations[self.panoptic_uuid]
-                    == int(g.object_id)
-                    + self._sim.habitat_config.instance_ids_start
-                )
-            return segmentation_sensor
-
-
-@registry.register_sensor
-class StartRecepSegmentationSensor(RecepSegmentationSensor):
-    cls_uuid: str = "start_recep_segmentation"
-
-    def _get_recep_goals(self, episode):
-        return episode.candidate_start_receps
-
-
-@registry.register_sensor
-class GoalRecepSegmentationSensor(RecepSegmentationSensor):
-    cls_uuid: str = "goal_recep_segmentation"
-
-    def _get_recep_goals(self, episode):
-        return episode.candidate_goal_receps
 
 
 class MultiObjSensor(PointGoalSensor):
