+++ habitat-lab/habitat/tasks/registration.py
@@ -9,6 +9,7 @@ from habitat.core.registry import registry
 from habitat.tasks.eqa import _try_register_eqa_task
 from habitat.tasks.nav import _try_register_nav_task
 from habitat.tasks.rearrange import _try_register_rearrange_task
+from habitat.tasks.ovmm import _try_register_ovmm_task
 from habitat.tasks.vln import _try_register_vln_task
 
 
@@ -26,3 +27,4 @@ _try_register_eqa_task()
 _try_register_nav_task()
 _try_register_vln_task()
 _try_register_rearrange_task()
+_try_register_ovmm_task()
